+++ src/PassPlanFilePlugin.groovy
@@ -3,12 +3,19 @@ import static TerraformEnvironmentStage.APPLY_COMMAND
 
 class PassPlanFilePlugin implements TerraformPlanCommandPlugin, TerraformApplyCommandPlugin, TerraformEnvironmentStagePlugin {
 
+    private static String directory = "./"
+
     public static void init() {
         PassPlanFilePlugin plugin = new PassPlanFilePlugin()
 
         TerraformEnvironmentStage.addPlugin(plugin)
         TerraformPlanCommand.addPlugin(plugin)
         TerraformApplyCommand.addPlugin(plugin)
+    }
+
+    public static withDirectory(String directory) {
+        PassPlanFilePlugin.directory = directory
+        return this
     }
 
     @Override
@@ -26,7 +33,7 @@ class PassPlanFilePlugin implements TerraformPlanCommandPlugin, TerraformApplyCo
     @Override
     public void apply(TerraformApplyCommand command) {
         String env = command.getEnvironment()
-        command.withDirectory("tfplan-" + env)
+        command.withPlanFile("tfplan-" + env)
     }
 
     public Closure stashPlan(String env) {
@@ -34,7 +41,9 @@ class PassPlanFilePlugin implements TerraformPlanCommandPlugin, TerraformApplyCo
             closure()
             String planFile = "tfplan-" + env
             echo "Stashing ${planFile} file"
-            stash name: planFile, includes: planFile
+            dir(directory) {
+                stash name: planFile, includes: planFile
+            }
         }
     }
 
@@ -42,9 +51,15 @@ class PassPlanFilePlugin implements TerraformPlanCommandPlugin, TerraformApplyCo
         return { closure ->
             String planFile = "tfplan-" + env
             echo "Unstashing ${planFile} file"
-            unstash planFile
+            dir(directory) {
+                unstash planFile
+            }
             closure()
         }
+    }
+
+    public static reset() {
+        directory = "./"
     }
 
 }